refactor(bench): split typecheck fixture runner
